organize path to create other donation list
