+++ frontend/src/pages/DonationList/CommonComponents/BottomMenu/BottomMenu.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/_commons";
+@import "../../../../styles/_commons";
 
 .containerMenu {
   background-color: #ffffff;
