+++ frontend/src/pages/DonationList/CommonComponents/DonationIsEmpty/DonationIsEmpty.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import './DonationIsEmpty.scss'
 
-import { donationListVoidText } from '../../../utils/strings'
-import { icEmptyDonationList } from '../../../assets/icons'
+import { donationListVoidText } from '../../../../utils/strings'
+import { icEmptyDonationList } from '../../../../assets/icons'
 
 const DonationIsEmpty = () => (
   <div className="containerIsEmpty">
