+++ frontend/src/pages/DonationList/CommonComponents/DonationIsEmpty/DonationIsEmpty.scss
