+++ frontend/src/pages/DonationList/CommonComponents/DonationItem/DonationItem.scss
