+++ frontend/src/pages/DonationList/CommonComponents/index.js
@@ -1,0 +1,4 @@
+export { BottomMenu } from './BottomMenu'
+export { DonationHeader } from './DonationHeader'
+export { DonationIsEmpty } from './DonationIsEmpty'
+export { DonationItem } from './DonationItem'
