+++ frontend/src/pages/DonationList/DonationListLeader/index.js
@@ -1,0 +1,3 @@
+import DonationListLeader from './DonationListLeader'
+
+export { DonationListLeader }
