+++ frontend/src/routes/index.js
@@ -9,14 +9,12 @@ import { Register } from '../pages/Register'
 import { Terms } from '../pages/Terms'
 import { Checklist } from '../pages/Checklist'
 
-import { DonationList } from '../pages/DonationList'
+import { DonationListLeader } from '../pages/DonationList'
 import { Donation } from '../pages/Donation'
 import { DonationProf } from '../pages/DonationProf'
 import { Received } from '../pages/Received'
 import { ReceivedCurrent } from '../pages/ReceivedCurrent'
 import { ReceivedCurrentProf } from '../pages/ReceivedCurrentProf'
-
-export { nextRouteToRole } from './routeWrapper'
 
 export const Routes = () => (
   <BrowserRouter>
@@ -26,8 +24,8 @@ export const Routes = () => (
       <Route path="/register" exact component={Register} />
       <Route path="/terms" exact component={Terms} isPrivate />
       <Route path="/checklist" exact component={Checklist} isPrivate />
-      <Route path="/donation-list" exact component={DonationList} isPrivate />
-      <Route path="/donation-list/admin" exact component={DonationList} isPrivate />
+      <Route path="/donation-list" exact component={DonationListLeader} isPrivate />
+      <Route path="/donation-list/admin" exact component={DonationListLeader} isPrivate />
       <Route path="/donation/:id/wait" exact component={Donation} isPrivate />
       <Route path="/donation/:id/wait/prof" exact component={DonationProf} isPrivate />
       <Route path="/donation/:id/received" exact component={Received} isPrivate />
